Update graphviz custom build

+++ GraphvizWrapper/include/gvplugin_render.h
@@ -15,6 +15,7 @@
 #include "types.h"
 #include "gvplugin.h"
 #include "gvcjob.h"
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -49,11 +50,11 @@ extern "C" {
 	void (*textspan) (GVJ_t * job, pointf p, textspan_t * span);
 	void (*resolve_color) (GVJ_t * job, gvcolor_t * color);
 	void (*ellipse) (GVJ_t * job, pointf * A, int filled);
-	void (*polygon) (GVJ_t * job, pointf * A, int n, int filled);
-	void (*beziercurve) (GVJ_t * job, pointf * A, int n, int);
-	void (*polyline) (GVJ_t * job, pointf * A, int n);
+	void (*polygon)(GVJ_t *job, pointf *A, size_t n, int filled);
+	void (*beziercurve)(GVJ_t *job, pointf *A, size_t n, int);
+	void (*polyline)(GVJ_t *job, pointf *A, size_t n);
 	void (*comment) (GVJ_t * job, char *comment);
-	void (*library_shape) (GVJ_t * job, char *name, pointf * A, int n, int filled);
+	void (*library_shape)(GVJ_t *job, char *name, pointf *A, size_t n, int filled);
     };
 
 #ifdef __cplusplus
